trivial: break types::state_store::state_value to sub-modules
